Add checkForSpace function
checkForSpace takes in one parameter - characterInput, which is taken
from the value of a button when it is clicked. checkForSpace checks if
either 1, the last character of currentString is an operator or 2, the
characterInput. If either of these conditions are met, a space is added
before the character is added to currentString.